meta-qti-bsp: startup-scripts: reset userfs to factory state
The userdata partition, i.e., the /data mountpoint,
is setup to be reformatted and all files therein are restored to factory state.
A reset is triggered from user space and takes effect in recovery cycle.
To trigger an userdata reset issue follwoing command:
   $ userfs-reset
and respond "y" at  prompt to proceed.

Change-Id: Ic499c46ae40df132ebe0202000ac11a0024e9b13