sdx20: find-recovery-partitions: fix some known issues
This CL has all of the below patches:

1) I320f32b6ee554844503bb13dbf1b1e42365e636a
Avoid ubiattch and mount of modem nand-partition

In case OTA is interrupted while modem is being upgraded,
device boots in recovery mode and the script tries to
ubiattach & mount modem partition again.
However, if modem upgrade was incomplete due to interruption in
previous run, then there exists no valid ubi-filesystem on the
partition. In this case, either ubiattach would fail or the volume-node
(i.e. /dev/ubix_y) will never be created and the script waits
indefinitely waiting for the same.

Avoid ubiattach/mount operation for modem partition. This is
feasible because modem is upgraded as raw partition with no
assumption about the filesystem that resides on it, thus avoiding need
of mounting the partition.

2) I98c19fb7afcc4547c6a7f1742039e502edbb2960
Wait and retry for ubi device and volume nodes

There is a possibility of attempting mount operation
on a ubi-volume node even before it is created. To make
sure this doesn't happen, send an additional argument to
FindAndAttachUBI(), which represents the number of ubi
volumes present on a ubi device.

After performing a ubiattach on mtd device:
  1) Explicitly check for ubi-device node i.e.
     /dev/ubiX to be created.
  2) Then check for each ubi volume node i.e.
     /dev/ubiX_Y to be created.

If any of the above nodes are not created within a timeout
period, reboot the device to exit recovery mode.
This will make sure that we don't do partial upgrade (of one
image but not another because ubi devices were not created).
Once booted into normal mode, one can always trigger OTA upgrade again.
Additionally, redirect all logging to /dev/kmsg explicitly.

3) I45e84a19eaaa6ad9e7d6c2855a650896df8457fe
Mount a tmpfs onto both /run and /var/volatile, else these
mount-points reside on rootfs and can fill it up.

4) Additionally, mount tmpfs on /res as well.
We dynamically write into /run/recovery_volume_detected
for recovery/updater to read. But /res is a mount point
on recoveryfs and hence if for some reason recoveryfs is
filled completely, this write fails and OTA aborts.
Mounting a tmpfs on /run reduces dependency on recoveryfs.

Change-Id: Id66a16095c170d1fa320b8d8c5340c974b49189b