early-init: to improve boot kpi for early init
boot kpi degrades by about 100ms after dm-verity is applied.
this can be improved by the following approaches.
1.put forward in time the mounting of rootfs and prefetch all
the early applications and so libraries they depend on. this
could improve boot time by 100-150ms.
2.put in correct order initial sequences for related regulators
and clocks that cpu-clock-msm8996 depends on and make cpu-clock-
msm8996 and msm-bus-fabric succeed and ready at the earliest time
possible. this is important since repeated failure and deferred
retries of these two components could block the initialization
of many drivers that depend on them. without this, the effort to
put forward in time the initialization of many driver modules
becomes useless.
3.try to parallelize more init functions in the system such as
camera and display drivers by running them in separate threads
and starting earlier in time.
4.set device probe that takes long time(>10ms) as asynchronous
if it's possible and allowed.
approach 2,3,4 will improve boot time by another about 100ms.
and total improved boot time is about 250ms on early-ethernet
and 190ms on early-cvbs.

Change-Id: I83b6a9d320e377a810f5623ef6eca5232bab24b5
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>