meta-qti-bsp: Add new mdev script to bring up mhi_dev_net0
Add a new mdev script that is called in mdev.conf to bring
the mhi_dev_net0 interface up. The reason for the script is
that we must check if /etc/resolv.conf is available before
calling ifup. There is a racing condition that makes it possible
for mdev to be called before the volatile links have been
created.

Change-Id: I20fba1e384f320a9c44712da572e01347793758f