kernel: defconfig: increase CONFIG_PANIC_TIMEOUT to 60
The original CONFIG_PANIC_TIMEOUT is 5s while the timeout
of watchdog is 10s in host. This makes host hard to
distinguish whether it's a normal reboot or a kernel
panic. Increasing CONFIG_PANIC_TIMEOUT to 60s will
always force host to detect a watchdog timeout and
collect ram-dump when gvm linux panics.

Change-Id: I6c784207ef831ac3fefdc632f3c24c3b0eadb017
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>