linux-msm: do not use kernel-yocto class
kernel-yocto is a complex class that is used to build the official
kernel tree maintained by the Yocto Project for the reference machines
(qemu, ...). this class does many things and makes assumptions on how
the kernel is released, especially how the metadata (defconfig) is
managed.

kernel-yocto class was introduced in this commit:
c1c53470ee7e (meta-qti-bsp: linux-msm: use kernel config fragment tool
to merge systemd config fragment)

the main reason was to be able to reuse the config fragment
mechanism. This patch removes the dependency on kernel-yocto class and
instead implements the config fragment support in
do_configure_prepend().

Change-Id: Iff38c85969c5430802fa6e19ae569fb287e13ac6
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>