mod-signing-keys: generate kernel signing keys separately
When kernel module signing is enabled (CONFIG_MODULE_SIG) a pair of
private and public keys is used. The public key is embedded in the
vmlinux image, and kernel modules are signed at build time using the
private key.

By default the kernel will create a pair of keys during do_compile
(see certs/Makefile). The keys are copied in STAGING_KERNEL_BUILDDIR,
so that they can be used to sign out of tree kernel
modules.

Everything works fine until TMPDIR is deleted! When that happens the
pair of keys is deleted permanently. When (re)building an out of tree
kernel module, there is a dependency on
virtual/kernel:do_shared_workdir, to ensure that
STAGING_KERNEL_BUILDDIR is rebuilt. In turns that means that we are
going to recompile the kernel recipe and generate a new pair of
keys. The new private key will be used to sign the kernel modules from
now on. However since the kernel recipe didn't change, we are using
the kernel artifact (vmlinux/zImage) from sstate, as such the kernel
image include the original public key, and the kernel modules won't
load.

So instead, we use mod-signing-keys to generate the pair of keys
ourselves, and keep them around (in sstate). Then we instruct the
kernel to use our pair of keys instead of generating them.

There is an issue in kernel.bbclass which was recently fixed in Poky
master branch, see:
1b5bd0bceb2b (kernel.bbclass: fix installation of modules signing certificates)

Now that we have mod-signing-keys recipe, we hit this issue. When
using external .pem file, the kernel will extract the public part of
the certificate (e.g. the .x509 file) from the .pem file. But in this
specific case kernel.bbclass will skip the copy of .x509 file, and we
can't sign any module. We can use this patch which implements the same
thing.

Change-Id: I171d91452746d06fced2187c66dfd804be6eb856
Signed-off-by: Nicolas Dechesne <ndeche@codeaurora.org>