linux-gvm-4.4: parallelize audio and pcie driver to improve bootkpi
Audio driver takes the most booting time, so we pick
out all audio dependency chain and run them in a
separate thread and bind it to a gold core. Pcie init
also takes some time and we can define it as
asynchronous to save more time. Finally, we postpone
the calling of wait_for_device_probe after the
mounting of rootfs so considerable rootfs mounting
time can be hidden in parallel with audio driver init.

Change-Id: I9f9237a45c5e1d2125c893a3e9deee6275baa0d9
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>