meta-qti-bsp: support dynamical memory hotplug
We need to support memory hotplug, but the range reserved
for hotplug changes for different ddr size. Calculating
range for each ddr size manually and configuring it in
cmdline will cause us to release separate kernel image for
each possible ddr size. This is not acceptable, so we need
to calculate it dynamically during booting.
In cmdline we tell linux which area is not supported for
memory hotplug and tell linux to support dynamical hotplug.
And we read the memory blocks that are reserved in sysfs
and hotplug them back in memory hotplug service.

Change-Id: I3eb6d876a6cb1f58d758e98ed8903ee8e1350eb6
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>