Use post-increment only in inffast.c
An old inffast.c optimization turns out to not be optimal anymore
with modern compilers, and furthermore was not compliant with the
C standard, for which decrementing a pointer before its allocated
memory is undefined. Per the recommendation of a security audit of
the zlib code by Trail of Bits and TrustInSoft, in support of the
Mozilla Foundation, this "optimization" was removed, in order to
avoid the possibility of undefined behavior.

Change-Id: I9332edef07cf9f0882a6b6291856a45debc8f1bb
Signed-off-by: mohaqure <mohaqure@codeaurora.org>