weston.ini:change default repaint-window value to 15
The original default repaint-window time is 7 ms, this is not enough
for multiple outputs repaint. Increase the value to optimize the
performance for multiple displays.
We used to change this value in source code of weston, it could be
bad for keeping pace with community. Therefore, change this value in
configure file.

Change-Id: I929c5539a710e308863a3c7b6c1bb4883570920f
Signed-off-by: Xiaoyu Huang <xiahua@codeaurora.org>