base-files: mount /tmp directly as a tmpfs
Before this change, /tmp is a soft link to /var/tmp, which is also
a soft link to /var/volatile/tmp. This can work well for most cases.
But here is one case where such implementaion is not good. It is
the set up where AppArmor(one kind of Linux Security Module) is used
to secure the system.
AppArmor considers a soft link file and its destination file as
seperated resources to give/deny access permission. So for some
application to get the access to a file like /tmp/<file-name>, the
permissions to /var/tmp and /var/volatile/tmp have to be explicitly
declared in its AppArmor profile in addition to the access to the
usual/standard /tmp directory. This could make the application's
AppArmor profile become platform-dependent.
So here, just mount /tmp as a tmpfs which is some usual implementation
like in some desktop Linux distribution.

Change-Id: Ie3cfc6277fe317e767a77287dd85a457fdbad501
Signed-off-by: Yong Ding <yongding@codeaurora.org>