linux-gvm-4.4: use async_schedule instead of workqueue
apr_init could wait for hab for several seconds in abnormal
cases and kernel will not wait for it to finish since it's
not a platform driver. So we use async_schedule instead of
original workqueue here since kernel will always wait for
all async jobs to finish before it moves forward and ends
the boot stage.

Change-Id: If8ae181ae9f33504d900f80c171980587fe0fae5
Signed-off-by: Jiacheng Zheng <jiaczhen@codeaurora.org>