Add entries for location and qwesd daemons inittab dynamically
Add entries for location and qwesd daemons for mdm9607 inittab
dynamically.
Since yocto branch is shared between different SIs some of which
have qwesd and location daemons enabled, some not. For those which
does not have, we need not add those entries in inittab. So add
for location and qwesd daemons for mdm9607 inittab dynamically.
Also remove the rmt_storage on this nand device SPs

Change-Id: Ife58c41b423dd3d29f31948c910fdd565b65c975
CRs-Fixed: 2901602