systemd: trigger login for early_init build
When enable early init, weston no longer trigger login operation,so XDG_RUNTIME_DIR won't be
created unless we adb shell/ssh to login to the board. this will cause issue when we try to
pass XDG_RUNTIME_DIR to container during bootup.
Trigger a login operations for this case.

Change-Id: Ibc0b4f1810ce8ae72f2325e4ffb7377648c3effa
Signed-off-by: Lei wang <leiwan@codeaurora.org>