evm: recovery: install ima/evm certificates for recoveryfs
The main system kernel treats unavailability of ima/evm
certificates as a security breach. The logic to trigger
fatal error (crash) in this case is embedded in the kernel
itself.

This commit ensures that if IMA/EVM is enabled in
build, the certificates are also installed in the
recovery file system to ensure kernel does not
crash. This is needed because recovery and main
system share same kernel image.

Change-Id: I846fd4d5d24152f2de0de0cfd6c9f862e2a1adbc
Signed-off-by: Rishi Gupta <rishgupt@codeaurora.org>