meta-qti-bsp: add extra dependency for veritybootimg
On a dm-verity enabled system, the normal make_bootimg function
will be executed first then the make_veritybootimg function will
be executed to append the verity command line.

But sometimes, in some build machine it is observed that first the
make_veritybootimg is executing and then the normal bootimg which
replaces the kernel command line and removes verity option.

Thus to ensure that always veritybootimg function should be executed
after system image and after normal bootimg, we add this dependency.

Change-Id: If333b9f2b78241bc49f8b6f7cce87ed9c891bc66
Signed-off-by: Pintu Kumar <quic_pintu@quicinc.com>