meta-qti-bsp: use patch function from base class
The local workspace synced with repo creates problems with the patching
in the kernel-yocto class using kgit-meta. This tries to apply the
patches using git and tracks the series in the .git directory.

As the kernel sources are copied into the yocto build directory the
links used by repo to refer to the location of the kernel repository
are broken. No valid git is found and the directory tree is travered
upwards to find the .git directory. In this case it usually will hit
the poky repository as it is next one valid. The patches are not
applied correctly in this case. Because of this problem other meta
layers trying to patch the kernel apply workarounds to get the sources
modified. Fixing the links to the repository can apply the patches
correctly but will follow the links and modify the original source
which is also not expected.

Skipping the patching through the kernel-yocto class and using the
patching done by the base class restores the common way to apply
source changes from recipes through the SRC_URI.

Change-Id: I8e3dd2cabf40578593333e3ff67d5fbeda289548