meta-qti-bsp: glib: fix CVE-2021-3800
libcharset: Drop a redundant environment variable

It was used for running tests when we built with autotools, but is no
longer used in the Meson build system. If we need something similar in
future, it should be done by adding internal API to override the
directory on a per-call basis, rather than loading a path from a shared
global table every time.

Change-Id: I7670c19e58fde2b589af9e9da8ac69cf665391a5