meta-qti-bsp: mark task merge_techpack_dtbos as nostamp
In the case when full build has been done once and then new techpack dtbo
is added, task merge_techpack_dtbos won't be triggered in the next build,
which is a not expected and as a consequence, the new added techpack dtbo
won't be present in the output image.
Add nostamp flag for the task so that merge_techpack_dtbos will be exectued
every time.

meanwhile, add correct dependency for merge_techpack_dtbos because it depends
on kernel_dtbs directory witch is generated by virtual/kernel:do_deploy

Change-Id: I69193b5ffe4d81bdf5ecf06653b378e648a38432