meta-qti-extra: Resolve USB ADB didn't work after resume from LPM
Hook script always assumed that the "/var/usb" directory exists,
but /var/usb didn't exist when (/vendor)/etc/usb is writable. The
usb mode will fail to save, which result in fail to restore in this
case. So try to create the path '/var/usb' manully when it not exists.

Change-Id: I8bffd5dfcf929238c2c2b7919725f9913ba4a32a
Signed-off-by: Fei Yang <quic_feiyan@quicinc.com>