Retain original layers when new layers are added to eSDK
Layers in external workspace used to replace respective
layers that are originally packed with eSDK if found common.
With this change new layers will be added as *-esdk layers.
Hence both new and old layers will be retained, and the new
layer will be prioritised in case of conflicting recipes.

Change-Id: I0ad70ee1b0bd7a279557ae817c4dfb41473c0401