CVE fixes for CVE-2021-33928, CVE-2021-33929, CVE-2021-33930
testcase_read: error out if repos are added or the system is changed too late

We must not add new solvables after the considered map was created, the solver
was created, or jobs were added. We may not changed the system after jobs have
been added.

(Jobs may point inside the whatproviedes array, so we must not invalidate this
area.)

Change-Id: I567f5a5cd540973cb4ec03ebba0914921cf95680
Signed-off-by: Ambika Pavan <quic_apavan@quicinc.com>