oot-dtbo: Avoid copying futile git data
Only copy the directories that are needed for serving
the git history, required for building the kernel.
Avoid copying any cache/optimization centric directory,
as it may not always exist on all systems, hence would
cause build failures in such cases. Moreover such
directories are not needed for the required functionality.

Change-Id: Ibc3aa3d84c55c3fbf77411052c8590ba51fd77c3
Signed-off-by: Sankalp Negi <quic_snegi@quicinc.com>