Revert "conf: machine: neo: Set 9128 as default composition for Neo"
This reverts commit 514e37a18fc10ade8187008c1e179b264f6289e7.

Reason for revert: It was set to 9128 before as team needed NCM
on bootup. But if NCM is the first interface, then the crashdump
mode wouldn't work. To avoid that revert back to 9060. When we have
a PID where NCM is present and Diag is the first interface, we will
shift to that new composition.

Change-Id: I503da435da0ed1ccc89cb4262cb5f6394ce9c9e4