meta-qti-bsp: Support XHCI unbind for the USB HOST controllers
Requirement for auto IVI targets is to support system suspend
even when a device is connected to the bus.

During suspend, the xhci hcd associated with dwc3 qcom can allow
the devices to program suspend via userspace.

There might be some devices which may not support autosuspend
and not allow power control for the device.

Also, there is a requirement to power off the GDSC for the controller
and hit RBSC. This results in losing power, resulting in loss of xhci
context.

Hence during resume, xhci cannot enable the already stored slot for the
connected devices before suspend.

Unbind the xhci instances before triggering suspend and bind the
xhci again resume to avert this problem.

Remove redundant power control options for the connected devices
in process.

Change-Id: I31232c91a9c663416fa77b5b177d1fb2e33c9e3f
Signed-off-by: Sriram Dash <quic_sriramd@quicinc.com>
Signed-off-by: Rajashekar kuruva <quic_kuruva@quicinc.com>