sa8797.inc: remove hgy-bbmask.inc
This file was originally used to fix SRC_URI checksum check issue in YP5.0.

However, each time we introduce a new recipe or update the componet in manifest, we have to update this file at the same time.

Now delete this file, will find new method to replace it later.

Change-Id: I321f3f7c1d3ae70322b859d45a6ab1eb93a89da3
Signed-off-by: pengy <quic_pengy@quicinc.com>