meta-qti-bsp: Disable zram on trustedvm-v3
When kswapd swaps out pages, PF_MEMALLOC is set, which results in zram
ignoring the min_watermark, and lowmem_reserve. This is problematic
because we intend for movable memory to stay in ZONE_MOVABLE, rather
than falling back to ZONE_NORMAL. If the later occurs to such a high
degree that all ZONE_NORMAL memory is consumed, as is observed to be
the case during a "C2PA" testcase, attempts to plug in additional
ZONE_MOVABLE memory may fail due to requiring a small amount of
GFP_KERNEL memory.

This increases idle TVM memory consumpion as follows:

After:
MemTotal:          83528 kB
~ # cat /proc/zoneinfo | grep managed
        managed  12690
        managed  8192

Before:
MemTotal:          67148 kB
~ # cat /proc/zoneinfo | grep managed
        managed  12691
        managed  4096

Change-Id: I42255bbd2e1847e910a742888b0b350e9c2452cf