qperf.bbclass: update PV for perf build
PR is dropped from WORKDIR in Yocto 5.0, so the workdir is the same for debug and perf build. This causes do_configure error "Argument list too long" for wlan recipe on perf build.

Append VARIANT (_perf) to PV so the WORKDIR is different for perf build to fix above error.

Change-Id: I02ccf7d5ed817fec203a5f5aae289418dc582ac8
Signed-off-by: xb <quic_xubaoj@quicinc.com>