meta-qti-bsp: update mount-copybind to create tmp copy then rename
In the event where power is lost during the copy operation, this
avoids the scenario where the destination file is created but the
copying of data is not completed. Since mount-copybind checks for
existence of file to determine if copy is required, such a failure
would be persistent. By creating a temporary copy first, the final
file only exists if the copy completes (since a rename is atomic).

Change-Id: I7f3d62a2481ee9a58e5d12f18cee3884df903ba0
(cherry picked from commit 8b0515ca1f3661074c1219a8613a1ea719a55d6a)