meta-qti-bsp: optimize the SRC_URI of linux-msm-headers_6.1.bb
Current SRC_URI is set to the msm-kernel path, build system will
vaidate the checksum for the whole msm-kernel code & copy kernle code
to the recipe workdir, which causes low build efficiency.

Recipe linux-msm-headers is used to provide kernel uapi headers which
is extracted from kernel-uapi-headers.tar.gz, and this recipe doesn't care the contents of msm-kernel, so set SRC_URI to the tarball itself
to address the build efficiency issue.

Change-Id: I6294b2dd7ee13bfdf0a6996184c922c66293cc37