meta-qti-base: Update oot-dtbo recipe for ddr size based dt overlay
In the device tree, certain nodes need to be updated based on the
DDR size available on the target. To achieve this, we can overlay
the device tree (dt) on top of dtb files generated for different
boards. We introduce OOT_DDR_DTBOS variable for the overlay device
tree created for various DDR sizes.

The device tree is selected by ABL based on the DDR type defined
in the board-id subtype value. Bits 8-10 in the board-id subtype
field are reserved for DDR type. This subtype value is updated
during the overlay process.

Below value are used for differnt ddr type
  DDRTYPE_8GB = 1
  DDRTYPE_12GB = 2
  DDRTYPE_16GB = 3
  DDRTYPE_24GB = 4
  DDRTYPE_36GB = 5
  DDRTYPE_48GB = 6
  DDRTYPE_64GB = 7

Change-Id: I74a6958d0e140c00d29e3cada397e35837a887c3
Signed-off-by: Jai Kumar Gautam <jgautam@qti.qualcomm.com>