system-prop: Add settings for COQOS Linux adb relay
The adb-relay is point to point adb protocol proxy. It is intended to be
a bridge between local (network socket) transport and USB transport. It
can be used to connect adb daemon running in virtual machine with
USB gadget device in order to make it available for adb server. Typical
application diagram looks looks next:

    +--VM2--+  +------VM1------+       +-------HOST------+
    |       |  |               |       |                 |
    | adbd--|--|---adb-relay--USB-----USB--adb-----adb   |
    |       |  |               |       |  server  client |
    +-------+  +---------------+       +-----------------+

So from the HOST point of view target looks like a normal USB device
with ADB Interface. However internally traffic is routed between virtual
machine with access to USB controller and virtual machine with actually
running adb daemon.

Change-Id: I8143076ddf541c983b403ba1f1915a34974351ce
Signed-off-by: Robert Michelsen <rmichels@qti.qualcomm.com>