meta-qti-base: Update meta for boot and bootloader image compilation of Nord LVGVM.
This mainly refers to LV043 bring up changes. Use "MACHINE:gvm-gen5" and "MACHINE_FEATURES:qtui-gvm" to distinguish betwwen the gvm-gen5 version and others in the meta.
To not change KP code, some stubs have to be made, such as lv dtb and efi from edk2. They are not correct, but can let the Yocto compile successfully.

Change-Id: Ia0df3a94441f9380d89244c1d6faceed1a0c30a8